Allow util to take objects
